Break out the `new ...` command hierarchy
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`new ...` hierarchy.
