+++ pkg/cmd/pulumi/newcmd/config_test.go
@@ -1,0 +1,304 @@
+// Copyright 2024, Pulumi Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package newcmd
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/config"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestParseConfigSuccess(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		Array    []string
+		Path     bool
+		Expected config.Map
+	}{
+		{
+			Array:    []string{},
+			Expected: config.Map{},
+		},
+		{
+			Array: []string{"my:testKey"},
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue(""),
+			},
+		},
+		{
+			Array: []string{"my:testKey="},
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue(""),
+			},
+		},
+		{
+			Array: []string{"my:testKey=testValue"},
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue("testValue"),
+			},
+		},
+		{
+			Array: []string{"my:testKey=test=Value"},
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue("test=Value"),
+			},
+		},
+		{
+			Array: []string{
+				"my:testKey=testValue",
+				"my:testKey=rewritten",
+			},
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue("rewritten"),
+			},
+		},
+		{
+			Array: []string{
+				"my:testKey=testValue",
+			},
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue("testValue"),
+			},
+		},
+		{
+			Array: []string{
+				"my:test.Key=testValue",
+			},
+			Expected: config.Map{
+				config.MustMakeKey("my", "test.Key"): config.NewValue("testValue"),
+			},
+		},
+		{
+			Array: []string{
+				"my:testKey=testValue",
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue("testValue"),
+			},
+		},
+		{
+			Array: []string{
+				"my:0=testValue",
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "0"): config.NewValue("testValue"),
+			},
+		},
+		{
+			Array: []string{
+				"my:true=testValue",
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "true"): config.NewValue("testValue"),
+			},
+		},
+		{
+			Array: []string{
+				`my:["test.Key"]=testValue`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "test.Key"): config.NewValue("testValue"),
+			},
+		},
+		{
+			Array: []string{
+				`my:outer.inner=value`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "outer"): config.NewObjectValue(`{"inner":"value"}`),
+			},
+		},
+		{
+			Array: []string{
+				`my:outer.inner.nested=value`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "outer"): config.NewObjectValue(`{"inner":{"nested":"value"}}`),
+			},
+		},
+		{
+			Array: []string{
+				`my:name[0]=value`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "name"): config.NewObjectValue(`["value"]`),
+			},
+		},
+		{
+			Array: []string{
+				`my:name[0][0]=value`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "name"): config.NewObjectValue(`[["value"]]`),
+			},
+		},
+		{
+			Array: []string{
+				`my:servers[0].name=foo`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "servers"): config.NewObjectValue(`[{"name":"foo"}]`),
+			},
+		},
+		{
+			Array: []string{
+				`my:testKey=false`,
+			},
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue("false"),
+			},
+		},
+		{
+			Array: []string{
+				`my:testKey=true`,
+			},
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue("true"),
+			},
+		},
+		{
+			Array: []string{
+				`my:testKey=10`,
+			},
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue("10"),
+			},
+		},
+		{
+			Array: []string{
+				`my:testKey=-1`,
+			},
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewValue("-1"),
+			},
+		},
+		{
+			Array: []string{
+				`my:testKey[0]=false`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewObjectValue(`[false]`),
+			},
+		},
+		{
+			Array: []string{
+				`my:testKey[0]=true`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewObjectValue(`[true]`),
+			},
+		},
+		{
+			Array: []string{
+				`my:testKey[0]=10`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewObjectValue(`[10]`),
+			},
+		},
+		{
+			Array: []string{
+				`my:testKey[0]=-1`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "testKey"): config.NewObjectValue(`[-1]`),
+			},
+		},
+		{
+			Array: []string{
+				`my:names[0]=a`,
+				`my:names[1]=b`,
+				`my:names[2]=c`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "names"): config.NewObjectValue(`["a","b","c"]`),
+			},
+		},
+		{
+			Array: []string{
+				`my:names[0]=a`,
+				`my:names[1]=b`,
+				`my:names[2]=c`,
+				`my:names[0]=rewritten`,
+			},
+			Path: true,
+			Expected: config.Map{
+				config.MustMakeKey("my", "names"): config.NewObjectValue(`["rewritten","b","c"]`),
+			},
+		},
+	}
+
+	//nolint:paralleltest // false positive because range var isn't used directly in t.Run(name) arg
+	for _, test := range tests {
+		test := test
+		t.Run(fmt.Sprintf("%v", test), func(t *testing.T) {
+			t.Parallel()
+
+			actual, err := ParseConfig(test.Array, test.Path)
+			assert.NoError(t, err)
+			assert.Equal(t, test.Expected, actual)
+		})
+	}
+}
+
+func TestSetFail(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		Array    []string
+		Expected config.Map
+	}{
+		{
+			Array: []string{`my:[""]=value`},
+		},
+		{
+			Array: []string{"my:[0]=value"},
+		},
+		{
+			Array: []string{`my:name[-1]=value`},
+		},
+		{
+			Array: []string{`my:key.secure=value`},
+		},
+	}
+
+	//nolint:paralleltest // false positive because range var isn't used directly in t.Run(name) arg
+	for _, test := range tests {
+		test := test
+		t.Run(fmt.Sprintf("%v", test), func(t *testing.T) {
+			t.Parallel()
+
+			_, err := ParseConfig(test.Array, true /*path*/)
+			assert.Error(t, err)
+		})
+	}
+}
