+++ pkg/cmd/pulumi/newcmd/new_test.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2023, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package newcmd
 
 import (
 	"context"
@@ -434,7 +434,7 @@ func TestInvalidTemplateName(t *testing.T) {
 		chdir(t, tempdir)
 
 		args := newArgs{
-			chooseTemplate:    chooseTemplate,
+			chooseTemplate:    ChooseTemplate,
 			interactive:       false,
 			yes:               true,
 			secretsProvider:   "default",
@@ -482,352 +482,6 @@ func TestInvalidTemplateName(t *testing.T) {
 		err := runNew(context.Background(), args)
 		assert.ErrorContains(t, err, "not found")
 	})
-}
-
-func TestParseConfigSuccess(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		Array    []string
-		Path     bool
-		Expected config.Map
-	}{
-		{
-			Array:    []string{},
-			Expected: config.Map{},
-		},
-		{
-			Array: []string{"my:testKey"},
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue(""),
-			},
-		},
-		{
-			Array: []string{"my:testKey="},
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue(""),
-			},
-		},
-		{
-			Array: []string{"my:testKey=testValue"},
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue("testValue"),
-			},
-		},
-		{
-			Array: []string{"my:testKey=test=Value"},
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue("test=Value"),
-			},
-		},
-		{
-			Array: []string{
-				"my:testKey=testValue",
-				"my:testKey=rewritten",
-			},
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue("rewritten"),
-			},
-		},
-		{
-			Array: []string{
-				"my:testKey=testValue",
-			},
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue("testValue"),
-			},
-		},
-		{
-			Array: []string{
-				"my:test.Key=testValue",
-			},
-			Expected: config.Map{
-				config.MustMakeKey("my", "test.Key"): config.NewValue("testValue"),
-			},
-		},
-		{
-			Array: []string{
-				"my:testKey=testValue",
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue("testValue"),
-			},
-		},
-		{
-			Array: []string{
-				"my:0=testValue",
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "0"): config.NewValue("testValue"),
-			},
-		},
-		{
-			Array: []string{
-				"my:true=testValue",
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "true"): config.NewValue("testValue"),
-			},
-		},
-		{
-			Array: []string{
-				`my:["test.Key"]=testValue`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "test.Key"): config.NewValue("testValue"),
-			},
-		},
-		{
-			Array: []string{
-				`my:outer.inner=value`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "outer"): config.NewObjectValue(`{"inner":"value"}`),
-			},
-		},
-		{
-			Array: []string{
-				`my:outer.inner.nested=value`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "outer"): config.NewObjectValue(`{"inner":{"nested":"value"}}`),
-			},
-		},
-		{
-			Array: []string{
-				`my:name[0]=value`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "name"): config.NewObjectValue(`["value"]`),
-			},
-		},
-		{
-			Array: []string{
-				`my:name[0][0]=value`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "name"): config.NewObjectValue(`[["value"]]`),
-			},
-		},
-		{
-			Array: []string{
-				`my:servers[0].name=foo`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "servers"): config.NewObjectValue(`[{"name":"foo"}]`),
-			},
-		},
-		{
-			Array: []string{
-				`my:testKey=false`,
-			},
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue("false"),
-			},
-		},
-		{
-			Array: []string{
-				`my:testKey=true`,
-			},
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue("true"),
-			},
-		},
-		{
-			Array: []string{
-				`my:testKey=10`,
-			},
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue("10"),
-			},
-		},
-		{
-			Array: []string{
-				`my:testKey=-1`,
-			},
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewValue("-1"),
-			},
-		},
-		{
-			Array: []string{
-				`my:testKey[0]=false`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewObjectValue(`[false]`),
-			},
-		},
-		{
-			Array: []string{
-				`my:testKey[0]=true`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewObjectValue(`[true]`),
-			},
-		},
-		{
-			Array: []string{
-				`my:testKey[0]=10`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewObjectValue(`[10]`),
-			},
-		},
-		{
-			Array: []string{
-				`my:testKey[0]=-1`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "testKey"): config.NewObjectValue(`[-1]`),
-			},
-		},
-		{
-			Array: []string{
-				`my:names[0]=a`,
-				`my:names[1]=b`,
-				`my:names[2]=c`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "names"): config.NewObjectValue(`["a","b","c"]`),
-			},
-		},
-		{
-			Array: []string{
-				`my:names[0]=a`,
-				`my:names[1]=b`,
-				`my:names[2]=c`,
-				`my:names[0]=rewritten`,
-			},
-			Path: true,
-			Expected: config.Map{
-				config.MustMakeKey("my", "names"): config.NewObjectValue(`["rewritten","b","c"]`),
-			},
-		},
-	}
-
-	//nolint:paralleltest // false positive because range var isn't used directly in t.Run(name) arg
-	for _, test := range tests {
-		test := test
-		t.Run(fmt.Sprintf("%v", test), func(t *testing.T) {
-			t.Parallel()
-
-			actual, err := parseConfig(test.Array, test.Path)
-			assert.NoError(t, err)
-			assert.Equal(t, test.Expected, actual)
-		})
-	}
-}
-
-func TestSetFail(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		Array    []string
-		Expected config.Map
-	}{
-		{
-			Array: []string{`my:[""]=value`},
-		},
-		{
-			Array: []string{"my:[0]=value"},
-		},
-		{
-			Array: []string{`my:name[-1]=value`},
-		},
-		{
-			Array: []string{`my:key.secure=value`},
-		},
-	}
-
-	//nolint:paralleltest // false positive because range var isn't used directly in t.Run(name) arg
-	for _, test := range tests {
-		test := test
-		t.Run(fmt.Sprintf("%v", test), func(t *testing.T) {
-			t.Parallel()
-
-			_, err := parseConfig(test.Array, true /*path*/)
-			assert.Error(t, err)
-		})
-	}
-}
-
-func TestErrorIfNotEmptyDirectory(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		desc  string
-		files []string
-		dirs  []string
-		ok    bool
-	}{
-		{
-			desc: "empty",
-			ok:   true,
-		},
-		{
-			desc:  "non-empty",
-			files: []string{"foo"},
-			dirs:  []string{"bar"},
-			ok:    false,
-		},
-		{
-			desc: "empty git repository",
-			dirs: []string{".git"},
-			ok:   true,
-		},
-		{
-			desc:  "non-empty git repository",
-			dirs:  []string{".git"},
-			files: []string{".gitignore"},
-			ok:    false,
-		},
-		{
-			desc: "every VCS",
-			dirs: []string{".git", ".hg", ".bzr"},
-			ok:   true,
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.desc, func(t *testing.T) {
-			t.Parallel()
-
-			path := t.TempDir()
-
-			// Fill test directory with files and directories
-			// requested by the test case.
-			for _, name := range tt.dirs {
-				err := os.MkdirAll(filepath.Join(path, name), 0o1700)
-				require.NoError(t, err)
-			}
-			for _, name := range tt.files {
-				err := os.WriteFile(filepath.Join(path, name), nil /* body */, 0o600)
-				require.NoError(t, err)
-			}
-
-			err := errorIfNotEmptyDirectory(path)
-			if tt.ok {
-				assert.NoError(t, err)
-			} else {
-				assert.Error(t, err)
-			}
-		})
-	}
 }
 
 func tempProjectDir(t *testing.T) string {
@@ -1129,29 +783,6 @@ func TestPulumiNewSetsTemplateTag(t *testing.T) {
 			tagsObject, ok := tagsValue.Value.(map[string]interface{})
 			assert.True(t, ok)
 			assert.Equal(t, tt.expected, tagsObject[apitype.ProjectTemplateTag])
-		})
-	}
-}
-
-func TestSanitizeTemplate(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		input    string
-		expected string
-	}{
-		{"https://user:pass@example.com/path?param=value", "https://example.com/path"},
-		{"https://user:pass@example.com", "https://example.com"},
-		{"https://example.com/path?param=value", "https://example.com/path"},
-		{"ssh://user@hostname/project/repo", "ssh://hostname/project/repo"},
-		{"typescript", "typescript"},
-		{"aws-typescript", "aws-typescript"},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.input, func(t *testing.T) {
-			t.Parallel()
-			result := sanitizeTemplate(tt.input)
-			assert.Equal(t, tt.expected, result)
 		})
 	}
 }
